webapp/.old/requirements.txt
# Requirements for Satellite Search
Django==4.0.4
djangorestframework==3.12.4
pytz==2022.1
psycopg2==2.9.2
django-cors-headers==3.10.0
tzdata==2022.1
requests==2.28.0
numpy==1.23.0
pyproj==3.3.1
scikit-learn==1.1.1
